ALPHA 1.8 fixed majority of connection bugs and made some balancing changes

+++ Scripts/Game/SpecificEffects/Effect126.gd
@@ -9,7 +9,7 @@ func _ready():
 	card = get_parent().get_parent()
 
 
-func Effect(attacker, _target, _n): # When attacked he dies and kill who attacked him
-	attacker.BoostByPos(attacker.Position, "health", -attacker.Health, attacker.Team) # kill the enemy
-	
-	card.BoostByPos(card.Position, "health", -card.Health, card.Team) # kill yourself
+func Effect(attacker, target, _n): # When attacked he dies and kill who attacked him
+	if card == target:
+		attacker.BoostByPos(attacker.Position, "health", -attacker.Health, attacker.Team) # kill the enemy
+		card.BoostByPos(card.Position, "health", -card.Health, card.Team) # kill yourself
